# api/openapi.yaml
openapi: 3.1.1
info:
  title: Go API Starter
  version: 1.0.0
  description: |
    This contract documents health, readiness, and the sample user CRUD routes.
    User routes are unauthenticated teaching examples, available when
    DB_ENABLED is true in every environment. Otherwise they return 404.
    Unknown paths return HTTP 404 using the NotFound error envelope defined
    under components.responses. CORS is disabled by default. When enabled,
    allowed preflight requests return 204
    before routing and denied preflight requests return 403; CORS response
    headers depend on the configured origin policy. Exact GET and HEAD requests
    to these health routes are exempt from rate limiting. Other routes,
    including docs, remain subject to configured limits. Docs routes are not
    API operations.
servers:
  - url: /
    description: Same origin as this API server
paths:
  /api/v1/users:
    get:
      operationId: listUsers
      summary: List users ordered by ID
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
      responses:
        '200':
          description: User list; an empty list is an empty JSON array.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/User'}
        '400': {$ref: '#/components/responses/UserError'}
        '404': {$ref: '#/components/responses/UserError'}
        '429': {$ref: '#/components/responses/UserError'}
        '500': {$ref: '#/components/responses/UserError'}
    post:
      operationId: createUser
      summary: Create a user
      requestBody: {$ref: '#/components/requestBodies/UserInput'}
      responses:
        '201':
          description: User created.
          headers:
            Location:
              schema: {type: string}
              description: Path of the created user.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '400': {$ref: '#/components/responses/UserError'}
        '404': {$ref: '#/components/responses/UserError'}
        '409': {$ref: '#/components/responses/UserError'}
        '413': {$ref: '#/components/responses/UserError'}
        '415': {$ref: '#/components/responses/UserError'}
        '429': {$ref: '#/components/responses/UserError'}
        '500': {$ref: '#/components/responses/UserError'}
  /api/v1/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer, format: int64, minimum: 1}
    get:
      operationId: getUser
      summary: Read a user
      responses:
        '200': {$ref: '#/components/responses/User'}
        '400': {$ref: '#/components/responses/UserError'}
        '404': {$ref: '#/components/responses/UserError'}
        '429': {$ref: '#/components/responses/UserError'}
        '500': {$ref: '#/components/responses/UserError'}
    put:
      operationId: updateUser
      summary: Replace both email and name of an existing user
      requestBody: {$ref: '#/components/requestBodies/UserInput'}
      responses:
        '200': {$ref: '#/components/responses/User'}
        '400': {$ref: '#/components/responses/UserError'}
        '404': {$ref: '#/components/responses/UserError'}
        '409': {$ref: '#/components/responses/UserError'}
        '413': {$ref: '#/components/responses/UserError'}
        '415': {$ref: '#/components/responses/UserError'}
        '429': {$ref: '#/components/responses/UserError'}
        '500': {$ref: '#/components/responses/UserError'}
    delete:
      operationId: deleteUser
      summary: Permanently delete a user
      responses:
        '204':
          description: User deleted; no response body.
        '400': {$ref: '#/components/responses/UserError'}
        '404': {$ref: '#/components/responses/UserError'}
        '429': {$ref: '#/components/responses/UserError'}
        '500': {$ref: '#/components/responses/UserError'}
  /health:
    get:
      operationId: getHealth
      summary: Check whether the API process is alive
      description: Liveness does not depend on database availability.
      responses:
        '200':
          description: The API process is alive.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
              example:
                status: ok
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    head:
      operationId: headHealth
      summary: Check API liveness without a response body
      description: Returns the GET status and headers with no response body.
      responses:
        '200':
          description: The API process is alive; the response has no body.
          headers:
            Content-Type:
              $ref: '#/components/headers/JSONContentType'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
        '405':
          description: The method is unsupported; no response body.
          headers:
            Content-Type:
              $ref: '#/components/headers/JSONContentType'
            Allow:
              $ref: '#/components/headers/AllowMethods'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
        '500':
          description: A handler panicked before committing its response; no response body.
          headers:
            Content-Type:
              $ref: '#/components/headers/JSONContentType'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
  /ready:
    get:
      operationId: getReadiness
      summary: Check whether the API can serve requests
      description: Readiness checks configured dependencies, including the database when enabled.
      responses:
        '200':
          description: All configured readiness checks passed.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
              example:
                status: ok
        '503':
          description: A configured readiness check failed.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessUnavailable'
              example:
                status: unavailable
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    head:
      operationId: headReadiness
      summary: Check readiness without a response body
      description: Returns the GET status and headers with no response body.
      responses:
        '200':
          description: All configured readiness checks passed; no response body.
          headers:
            Content-Type:
              $ref: '#/components/headers/JSONContentType'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
        '503':
          description: A configured readiness check failed; no response body.
          headers:
            Content-Type:
              $ref: '#/components/headers/JSONContentType'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
        '405':
          description: The method is unsupported; no response body.
          headers:
            Content-Type:
              $ref: '#/components/headers/JSONContentType'
            Allow:
              $ref: '#/components/headers/AllowMethods'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
        '500':
          description: A handler panicked before committing its response; no response body.
          headers:
            Content-Type:
              $ref: '#/components/headers/JSONContentType'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
components:
  requestBodies:
    UserInput:
      required: true
      description: Both fields are required. Body limit is 4096 bytes. Email is trimmed and lowercased; name is trimmed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/UserInput'}
  headers:
    JSONContentType:
      description: The response media type.
      schema:
        type: string
        const: application/json
    AllowMethods:
      description: Methods supported by this path.
      schema:
        type: string
        const: GET, HEAD
    ContentTypeOptions:
      description: Prevents MIME type sniffing.
      schema:
        type: string
        const: nosniff
  schemas:
    User:
      type: object
      required: [id, email, name]
      additionalProperties: false
      properties:
        id: {type: integer, format: int64, minimum: 1}
        email: {type: string, maxLength: 254}
        name: {type: string, minLength: 1, maxLength: 100}
    UserInput:
      type: object
      required: [email, name]
      additionalProperties: false
      properties:
        email:
          type: string
          description: A single bare email address, at most 254 bytes after trimming; must be unique ignoring case.
        name:
          type: string
          description: Nonempty and at most 100 Unicode characters after trimming.
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          const: ok
      additionalProperties: false
    ReadinessUnavailable:
      type: object
      required: [status]
      properties:
        status:
          type: string
          const: unavailable
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [not_found, method_not_allowed, internal_error, invalid_input, email_taken, body_too_large, unsupported_media_type, rate_limited]
            message:
              type: string
          additionalProperties: false
      additionalProperties: false
  responses:
    User:
      description: A user with exactly id, email and name.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/User'}
    UserError:
      description: Invalid input (400), unavailable route or missing user (404), duplicate email (409), body too large (413), unsupported media type (415), rate limit (429), or internal failure (500).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    NotFound:
      description: The request path does not match a registered route.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: not_found
              message: Resource not found
    MethodNotAllowed:
      description: The method is unsupported. The Allow header lists GET and HEAD.
      headers:
        Allow:
          description: Methods supported by this path.
          $ref: '#/components/headers/AllowMethods'
        X-Content-Type-Options:
          $ref: '#/components/headers/ContentTypeOptions'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: method_not_allowed
              message: Method not allowed
    InternalServerError:
      description: A handler panicked before committing its response.
      headers:
        X-Content-Type-Options:
          $ref: '#/components/headers/ContentTypeOptions'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: internal_error
              message: Internal server error